requirements.txt
asgiref==3.7.2
bleach==3.3.0
dj-database-url==2.1.0
Django==5.0.1
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-js-asset==2.2.0
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
webencodings==0.5.1
